Args

Description

Language : Haskell2010

Argument parser generated by arggen v1.0.0 at 2019-09-10 19:57:26.817940 (c) Edward Jones 2019.

Documentation

parseArgs :: String -> [String] -> Args

For a named program, parse the argument list

parseArgv :: IO Args

Parse the command-line arguments

data Args

Record to hold parsed arguments

input :: Args -> String

Get value from arguments: "Input identifier (how this is used depends on the operation performed)"

includeLocation :: Args -> Bool

Get value from arguments: "Output the emperor header location"

addDependency :: Args -> String

Get value from arguments: "Add the dependency specified by the -i flag to the current package"

installDependencies :: Args -> Bool

Get value from arguments: "Install the dependencies of the current project"

cFlags :: Args -> Bool

Get value from arguments: "Output GCC flags"

libs :: Args -> Bool

Get value from arguments: "Output required library flags"

binaryInstallLocation :: Args -> Bool

Get value from arguments: "Location where binary executables are located once installed"

libraryInstallLocation :: Args -> Bool

Get value from arguments: "Location where emperor libraries are located once installed"

dataInstallLocation :: Args -> Bool

Get value from arguments: "Location where data files for emperor programs may be installed (read/write)"

force :: Args -> Bool

Get value from arguments: "Force updates regardless of whether a dependency is already present"

dryRun :: Args -> Bool

Get value from arguments: "Perform a dry run (print the commands which would be executed without running them)"

entryPoint :: Args -> Bool

Get value from arguments: "Enforce that this is the entry-point"

updatePackageRepo :: Args -> Bool

Get value from arguments: "Update the list of known packages"

languageHeaderLocation :: Args -> Bool

Get value from arguments: "Output the location of the location of the language-specific C headers. These are only used by the *emperor* compiler, any headers provided by packages should be located in the directory specified by the --include-location flag, instead of -#."

getPackageLocation :: Args -> String

Get value from arguments: "Output the location of a particular package"

bareProject :: Args -> Bool

Get value from arguments: "Output compilation flags without libraries or include locations. Useful to ensure that foreign C code is compatible with emperor code. This also makes -lb do nothing."