Formatter.Formatter

Description

Language : Haskell2010

This file defines the standard format for all Emperor programs.

Documentation

format :: Format a => FormatContext -> a -> String

Create a string which represents the object in the current context

formatFresh :: Format a => a -> String

Format from a blank context

class Format a where

Describes items which may be formatted

Minimal complete definition

format

Methods

format :: FormatContext -> a -> String

Create a string which represents the object in the current context

formatFresh :: a -> String

Format from a blank context

type FormatContext = Int

The information required to format code in a given context