Install

Description

Language : Haskell2010

This installs dependencies of a package and where necessary will download a list of the available packages.

Documentation

doInstallDependencies :: Args -> IO ()

Install the dependencies required by the command-line arguments

ensurePackageRepoExists :: Args -> IO ()

Installs the package repository if necessary

installPackageDependencies :: Args -> Package -> IO ()

Install the dependencies of a given package