Logger.Logger

Description

Language : Haskell2010

Used to create loggers which output to stderr using ANSI formatted output.

Documentation

makeLoggers :: Args -> IO (Logger, Logger, Logger, Logger)

Make a set of loggers according to the command-line argument. (Specifically whether to print verbose output and whether to override the heuristic-based method for checking colour-compatibility.)

type Loggers = (Logger, Logger, Logger, Logger)

Defines a set of logger functions. From left to right these are: errorLogger, infoLogger, successLogger, warningLogger

type Logger = String -> IO ()

Type of a logger function which writes to some output