Package

Description

Language : Haskell2010

Defines data types to represent and manipulate emperor packages. Defines functions to translate packages in to JSON and back.

Documentation

data Package

Defines an emperor package

Constructors

Package

Fields

packageName :: String

author :: Author

dependencies :: [Dependency]

packageVersion :: String

files :: [String]

license :: String

data Author

An author of an emperor package

Constructors

Author

Fields

authorName :: String

email :: String

url :: String

data Dependency

A dependency of a given package

Constructors

Dependency

Fields

dependencyName :: String

dependencyVersion :: String

class HasName a t | a -> t where

Defines data items which have a name

Minimal complete definition

name

Methods

name :: a -> t

Obtain the name of a specified data item

class HasVersion a t | a -> t where

Defines data items which have a version

Minimal complete definition

version

Methods

version :: a -> t

Obtain the version of a specified data item

hasDependency :: Package -> Dependency -> Bool

Returns whether a given package depends on some version of a given dependency

insertDependency :: Package -> Dependency -> Package

Adds a dependency to the list associated with a package

parseDependencyString :: String -> (String, Maybe String)

Parse a dependency string in to a package name and maybe a version

version :: HasVersion a t => a -> t

Obtain the version of a specified data item

name :: HasName a t => a -> t

Obtain the name of a specified data item

getPackage :: FilePath -> IO (Maybe Package)

Get the package specified by a given manifest

getPackageFromDirectory :: FilePath -> IO (Maybe Package)

Obtain the package located at a given directory

getPackageMeta :: Args -> IO (Maybe Package)

Obtain the package given by the command-line arguments