Parser.AST

Description

Language : Haskell2010

This module defines the data-types used to form the abstract syntax tree of the emperor language.

Documentation

data Assignment

Data-structure to represent an assignment statement

Constructors

Assignment (Maybe EmperorType) Ident Expr AlexPosn

data AST

Data type to represent the abstract syntax tree for a single module. This is specified by its name, its imports and its code.

Constructors

AST ModuleHeader [Import] [ModuleItem]

data BodyBlock

Represents a single construction in the body of a function. This may be a further construct or just a single line

Constructors

Line BodyLine AlexPosn
A single line of code

IfElse Expr [BodyBlock] [BodyBlock] AlexPosn
An if-else block

While Expr [BodyBlock] AlexPosn
A while-loop

For Ident Expr [BodyBlock] AlexPosn
A for-loop

Repeat Expr [BodyBlock] AlexPosn
A repeat-loop

With Assignment [BodyBlock] AlexPosn
A resource acquisition

Switch Expr [SwitchCase] AlexPosn
A switch-case statement

data BodyLine

Data-structure for a single body-line

Constructors

AssignmentC Assignment

QueueC Queue

CallC Call

Return (Maybe Expr) AlexPosn

data Call

Represents the use of a function

Constructors

Call Purity Ident [Expr] AlexPosn

data Expr

Data-structure to represent an expression

Constructors

Value Value AlexPosn

Neg Expr AlexPosn

Add Expr Expr AlexPosn

Subtract Expr Expr AlexPosn

Multiply Expr Expr AlexPosn

Divide Expr Expr AlexPosn

Modulo Expr Expr AlexPosn

Less Expr Expr AlexPosn

LessOrEqual Expr Expr AlexPosn

Greater Expr Expr AlexPosn

GreaterOrEqual Expr Expr AlexPosn

Equal Expr Expr AlexPosn

NotEqual Expr Expr AlexPosn

Not Expr AlexPosn

AndStrict Expr Expr AlexPosn

AndLazy Expr Expr AlexPosn

OrStrict Expr Expr AlexPosn

OrLazy Expr Expr AlexPosn

Implies Expr Expr AlexPosn

Xor Expr Expr AlexPosn

ShiftLeft Expr Expr AlexPosn

ShiftRight Expr Expr AlexPosn

ShiftRightSameSign Expr Expr AlexPosn

Set [Expr] AlexPosn

Tuple [Expr] AlexPosn

List [Expr] AlexPosn

data FunctionDef

Describes the definition of a function

Constructors

FunctionDef FunctionTypeDef [Ident] [BodyBlock] AlexPosn

data FunctionTypeDef

Describes the definition of the type of a function

Constructors

FunctionTypeDef Ident EmperorType AlexPosn

data Ident

Data-structure to represent an identifier

Constructors

Ident String AlexPosn

data Import

A single imported file

Constructors

Import ImportLocation (Maybe [Ident]) AlexPosn

data ImportLocation

Location of an import and how to treat it

Constructors

ImportLocation ImportType Ident AlexPosn

data ImportType

The type of an import

Constructors

Local
Indicates a file in the current project

Global
Indicates a file in the global installation

data ModuleHeader

A single module header

Constructors

Module Ident AlexPosn

data ModuleItem

Describes a single named item in the module

Constructors

Component Ident (Maybe [TypeComparison]) [FunctionDef] AlexPosn

TypeClass Ident (Maybe [TypeComparison]) [FunctionTypeDef] AlexPosn

FunctionItem FunctionDef AlexPosn

data Queue

Data-structure to represent an queue statement

Constructors

Queue (Maybe EmperorType) Ident Expr AlexPosn

data SwitchCase

Data-struecture for the switch-case statement

Constructors

SwitchCase Expr BodyBlock AlexPosn

data TypeComparison

Describes an explicit type assertion

Constructors

IsSubType Ident AlexPosn

IsSubTypeWithImplementor Ident Ident AlexPosn

data Value

Data-structure to represent a single value

Constructors

IDC AlexPosn

Integer Integer AlexPosn

Real Double AlexPosn

Char Char AlexPosn

StringV String AlexPosn

IdentV Ident AlexPosn

Bool Bool AlexPosn

CallV Call AlexPosn