Parser.EmperorLexer

Description

Language : Haskell2010

This module defines the machinery to lexically analyse the Emperor language given in an input string.

Documentation

data Alex a

State information for an Alex lexer

data AlexPosn

Constructor to represent the position of a token in code. Format: current line, current character on line, current total character over all lines

Constructors

AlexPn !Int !Int !Int

data Token

Type to represent tokens in the output stream

Constructors

TDocAssignmentLine
Documentation field line

Fields

position :: AlexPosn

TDocLine
Generic line of documentation

Fields

content :: String

position :: AlexPosn

TInteger
An integral literal

Fields

intVal :: Integer

position :: AlexPosn

TBool
A boolean literal

Fields

isTrue :: Bool

position :: AlexPosn

TReal
A real/floating-point literal

Fields

realVal :: Double

position :: AlexPosn

TChar
A single character literal

Fields

charVal :: Char

position :: AlexPosn

TString
A single character literal

Fields

stringVal :: String

position :: AlexPosn

TIf
Keyword: if

Fields

position :: AlexPosn

TElse
Keyword: else

Fields

position :: AlexPosn

TWhile
Keyword: while

Fields

position :: AlexPosn

TRepeat
Keyword: repeat

Fields

position :: AlexPosn

TWith
Keyword: with

Fields

position :: AlexPosn

TSwitch
Keyword: switch

Fields

position :: AlexPosn

TFor
Keyword: for

Fields

position :: AlexPosn

TImport
Keyword: import

Fields

position :: AlexPosn

TModule
Keyword: module

Fields

position :: AlexPosn

TIdent
An identifier

Fields

identifierVal :: String

position :: AlexPosn

TPartSeparator
;

Fields

position :: AlexPosn

TBlockSeparator
#

Fields

position :: AlexPosn

TQueue
<-

Fields

position :: AlexPosn

TGoesTo
->

Fields

position :: AlexPosn

TGets
=

Fields

position :: AlexPosn

TLParenth
(

Fields

position :: AlexPosn

TRParenth
)

Fields

position :: AlexPosn

TLBracket
[

Fields

position :: AlexPosn

TRBracket
]

Fields

position :: AlexPosn

TLBrace
{

Fields

position :: AlexPosn

TRBrace
}

Fields

position :: AlexPosn

TImpure
@

Fields

position :: AlexPosn

TPlus
+

Fields

position :: AlexPosn

TMinus
-

Fields

position :: AlexPosn

TDivide
/

Fields

position :: AlexPosn

TModulo
%

Fields

position :: AlexPosn

TTimes
*

Fields

position :: AlexPosn

TShiftLeft
<<

Fields

position :: AlexPosn

TShiftRight
>>

Fields

position :: AlexPosn

TShiftRightSameSign
>>>

Fields

position :: AlexPosn

TAndScrict
&

Fields

position :: AlexPosn

TAndLazy
&&

Fields

position :: AlexPosn

TOrStrict
|

Fields

position :: AlexPosn

TOrLazy
||

Fields

position :: AlexPosn

TNot
!

Fields

position :: AlexPosn

TXor
^

Fields

position :: AlexPosn

TLessThan
<

Fields

position :: AlexPosn

TLessThanOrEqual
<=

Fields

position :: AlexPosn

TGreaterThan
>

Fields

position :: AlexPosn

TGreaterThanOrEqual
>=

Fields

position :: AlexPosn

TImplies
=>

Fields

position :: AlexPosn

TEqual
==

Fields

position :: AlexPosn

TNotEqual
!=

Fields

position :: AlexPosn

TComma
,

Fields

position :: AlexPosn

TIntT
int

Fields

position :: AlexPosn

TBoolT
bool

Fields

position :: AlexPosn

TRealT
real

Fields

position :: AlexPosn

TCharT
char

Fields

position :: AlexPosn

TStringT
char

Fields

position :: AlexPosn

TUnit
()

Fields

position :: AlexPosn

TAnyT
AnyT

Fields

position :: AlexPosn

TColon
AnyT

Fields

position :: AlexPosn

TIsSubType
<:

Fields

position :: AlexPosn

TIsImplementeBy
<~

Fields

position :: AlexPosn

TIsType
::

Fields

position :: AlexPosn

TClass
class

Fields

position :: AlexPosn

TComponent
component

Fields

position :: AlexPosn

TIDC
_

Fields

position :: AlexPosn

TReturn
_ | TEoL { position :: AlexPosn } -- ^ \n

Fields

position :: AlexPosn

TEoF
\0

lexWrap :: (Token -> Alex a) -> Alex a

Wrapper function for the lexer---allows the monadic lexer to be used with a monadic parser

alexError :: String -> Alex a

Function to return an Alex error state

runAlex :: String -> Alex a -> Either String a

Function to run the generated lexer