Parser.EmperorParser

Description

Language : Haskell2010

This module defines the machinery to parse the Emperor language from a token stream generated by the Emperor lexer.

Documentation

parseEmperor :: Alex AST

Run the emperor parser

parseREPL :: Alex ModuleItem

Run the parser for REPL lines