Parser.EmperorParserWrapper

Description

Language : Haskell2010

This forms a wrapper to make the Emperor parser and lexer easier to use. It extends functionality by allowing the format of strings, arbitrary files on disk, and stdin when the file "-" is used.

Documentation

data AST

Data type to represent the abstract syntax tree for a single module. This is specified by its name, its imports and its code.

parse :: String -> IO (Either String AST)

Parse a file of "-" for stdin

parseFile :: FilePath -> IO (Either String AST)

Parse an arbitrary file on disk

parseString :: String -> Either String AST

Parse a given string