Types.Checker

Description

Language : Haskell2010

This module defines type-checking for Emperor.

Documentation

class TypeCheck a where

Describes objects which may be type-checked.

Minimal complete definition

(>-)

Methods

(>-) :: TypeEnvironment -> a -> TypeCheckResult infix 2

Check the assertion that a given a has valid type

(>-) :: TypeCheck a => TypeEnvironment -> a -> TypeCheckResult infix 2

Check the assertion that a given a has valid type