Types.Environment

Description

Language : Haskell2010

This module defines the type environments of Emperor programs.

Documentation

(=>>) :: TypeEnvironment -> String -> TypeJudgementResult

Get a value from a type environment

filterEnvironment :: (String -> Bool) -> TypeEnvironment -> TypeEnvironment

Filter the entries of the type environment

fromList :: [(String, EmperorType)] -> TypeEnvironment

Create a type environment from a list

has :: TypeEnvironment -> String -> Bool

Check whether a given type environment contains a symbol of the given name

newTypeEnvironment :: TypeEnvironment

Creates a fresh type-environment

newtype TypeEnvironment

An environment which maps names to types

Constructors

TypeEnvironment [(String, EmperorType)]

insert :: String -> EmperorType -> TypeEnvironment -> TypeEnvironment

Add a type assertion to the environment

unsafeGet :: String -> TypeEnvironment -> EmperorType

Get a value from a type environment under the assertion that it already exists. This should only be used for types guaranteed to be in the prelude.