Types.Imports.Imports

Description

Language : Haskell2010

From a given file, returns the type environment specified by its interface

Documentation

getEnvironment :: Loggers -> [Import] -> IO (Either String TypeEnvironment)

Given a set of imports, obtain the type environment they form.

getLocalEnvironment :: AST -> Either String TypeEnvironment

Obtain the type environment created by the content of the module

writeHeader :: FilePath -> AST -> IO ()

Write a header to a file