Types.Imports.JsonIO

Description

Language : Haskell2010

This defines the machinery to encode and decode the type environments exported by modules in JSON. Common I/O operations are also provided

Documentation

data Header

Defines module header file contents

Constructors

Header Ident [ImportLocation] TypeEnvironment

isHeaderFile :: Loggers -> FilePath -> IO Bool

Checks whether a given header file exists

readHeader :: Loggers -> FilePath -> IO (Either String Header)

Read a header from a file, obtain its type environment

writeHeader :: Header -> FilePath -> IO ()

Write a type environment to a file to make a header