Types.Judger

Description

Language : Haskell2010

This module defines typing judgements for the AST. Types are found and checked for compatibility.

Documentation

(|>) :: Typable a => TypeEnvironment -> a -> TypeJudgementResult infixl 2

Judge the type of a given construct under a particular typing environment.

judge :: Typable a => a -> TypeJudgementResult

Obtain the type of a given construct from a fresh environment.

class Typable a where

Class describing constructs which may be assigned a type.

Minimal complete definition

(|>)

Methods

judge :: a -> TypeJudgementResult

Obtain the type of a given construct from a fresh environment.

(|>) :: TypeEnvironment -> a -> TypeJudgementResult infixl 2

Judge the type of a given construct under a particular typing environment.