Types.Results

Description

Language : Haskell2010

This module defines valid type structures and the results which may be returned by the argument checker or resolver.

Documentation

data EmperorType

Data to represent all Emperor types

Constructors

IntP
Integer primitive

CharP
Character primitive

BoolP
Boolean primitive

RealP
Real number primitive

ESet EmperorType
Set composite

EList EmperorType
List composite

ETuple [EmperorType]
Tuple composite

ERecord (Map String EmperorType)
Record composite

EFunction Purity EmperorType EmperorType
Function composite

Any
Universal super-type

Unit
Universal sub-type

data TypeCheckResult

The result of a type-check; this indicates a valid or invalid relation. Where this is invalid, a message is given.

Constructors

Pass
Indicates a correct type statement

Fail String
Indicates an invalid type and provides a reason

data TypeJudgementResult

The result of a typing judgement. This is either an error indicating a problem or a type

Constructors

Valid EmperorType
Valid type

Invalid String
Indicates an incorrect type and gives an explanation

getTypeList :: EmperorType -> [EmperorType]

Get a list of the types mentioned in a function type declaration. Example: getTypeList a -> b -> c = [a,b,c]

isValid :: TypeOp a => a -> Bool

Indicates whether a type operation has returned a valid result

isValidAnd :: EmperorType -> TypeJudgementResult -> Bool

Check that a type judgement result is valid and is equal to a given type

data Purity

Marker for whether a function is pure or impure

Constructors

Pure

Impure

class TypeOp a where

Class of types which describe type operation results

Minimal complete definition

isValid

Methods

isValid :: a -> Bool

Indicates whether a type operation has returned a valid result

unpackTypes :: [TypeJudgementResult] -> Either String [EmperorType]

Convert a list of type judgements in to a list of emperor types or an error.