Types.Types

Description

Language : Haskell2010

This module enables the type checking and judgement modules to be called more easily.

Documentation

resolveTypes :: Loggers -> AST -> IO TypeCheckResult

Find any problems with the typing of results and obtain types if no problems are found.

data TypeCheckResult

The result of a type-check; this indicates a valid or invalid relation. Where this is invalid, a message is given.

Constructors

Pass
Indicates a correct type statement

Fail String
Indicates an invalid type and provides a reason

writeHeader :: FilePath -> AST -> IO ()

Write a header to a file